♻️ Move Event, EventOptions, StepOptions and etc to opts and add temporary alias to sqlslog package
